auth models fixed
